v1.1.14 -- fix an important bug in calculatePvalues()
This commit includes tests so that all functions in derfinder are now tested.
Thus, in the future it will be easier to catch any bug introduced by
changing the code.

What prompted us to do this was a bug identified by reproducing results from
version 0.0.63. The cause of the bug ended up being an incorrectly set value
for the maxRegionGap parameter inside calculatePvalues(). It was set to 300
instead of 0, meaning that the segments of base pairs passing the data filter
were incorrectly grouped as findRegions() did have the correct default value of
0.

If you have results from prior to this fix, you could fix your regions using
https://gist.github.com/bf85e2c7d5d1f8197707 However, note that the p-values
will be approximated as there is no way to fix the null areas. If you want the
correct p-values the only option is to re-run derfinder.